Replace startIfNecessary with start method
The "start if necessary" strategy makes debugging very hard.
There are two possible error conditions which are silently hidden:

1. A new ImageJ server couldn't be started because an existing one is still
   running.
2. The existing ImageJ server surprisingly stopped.

In both cases the user does NOT get's no notification about the problems.
And the problems that arise from that, for example wrong plugin version
executed, will seem complete mysterious to the user.