Broken Example with DiskCachedCellImg
The simple example RotateSingleDataset cannot be executed in the
CellLoader of a DiskCachedCellImg.

A come across the same problem when I want to use scijava-parallel with
Labkit. In Labkit the segmentation is performed in the CellLoader of
a DiskCachedCellImage. This segmentation should be executed on the cluster
but somehow scijava-parallel fails inside the CellLoader of a
DiskCachedCellImg.